+++ tomcat/src/main/java/org/apache/coyote/handler/Handler.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.handler;
 
 import java.io.IOException;
-import org.apache.coyote.HttpRequest;
-import org.apache.coyote.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public interface Handler {
